# docs/swagger.yaml
definitions:
...
  inout.BaseResponse:
    properties:
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  inout.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  model.OrganizationStatus:
    enum:
    - active
    - suspended
    - cancelled
    type: string
    x-enum-varnames:
    - OrganizationStatusActive
    - OrganizationStatusSuspended
    - OrganizationStatusCancelled
  model.PlanType:
    enum:
    - free
    - starter
    - professional
    - enterprise
    type: string
    x-enum-varnames:
    - PlanTypeFree
    - PlanTypeStarter
    - PlanTypeProfessional
    - PlanTypeEnterprise
...
  organization.CreateOrganizationRequest:
    properties:
      description:
        type: string
      logo_url:
        type: string
      max_projects:
        type: integer
      max_users:
        type: integer
      name:
        maxLength: 200
        minLength: 2
        type: string
      plan_type:
        $ref: '#/definitions/model.PlanType'
      slug:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  organization.InviteMemberRequest:
    properties:
      email:
        type: string
      role:
        enum:
        - member
        - admin
        type: string
    required:
    - email
    - role
    type: object
  organization.InviteOut:
    properties:
      data:
        $ref: '#/definitions/organization.InviteResult'
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  organization.InviteResult:
    properties:
      email:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  organization.Member:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      joined_at:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  organization.MembersOut:
    properties:
      data:
        items:
          $ref: '#/definitions/organization.Member'
        type: array
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  organization.Organization:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      logo_url:
        type: string
      max_projects:
        type: integer
      max_users:
        type: integer
      name:
        type: string
      plan_type:
        $ref: '#/definitions/model.PlanType'
      slug:
        type: string
      status:
        $ref: '#/definitions/model.OrganizationStatus'
      updated_at:
        type: string
    type: object
  organization.OrganizationListOut:
    properties:
      error_code:
        type: integer
      error_description:
        type: string
      list:
        items:
          $ref: '#/definitions/organization.Organization'
        type: array
      meta:
        $ref: '#/definitions/inout.PaginationMeta'
    type: object
  organization.OrganizationOut:
    properties:
      data:
        $ref: '#/definitions/organization.Organization'
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  organization.PendingInvite:
    properties:
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      invited_at:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
  organization.PendingInvitesOut:
    properties:
      data:
        items:
          $ref: '#/definitions/organization.PendingInvite'
        type: array
      error_code:
        type: integer
      error_description:
        type: string
    type: object
  organization.UpdateMemberRoleRequest:
    properties:
      role:
        enum:
        - member
        - admin
        type: string
    required:
    - role
    type: object
  organization.UpdateOrganizationRequest:
    properties:
      description:
        type: string
      logo_url:
        type: string
      max_projects:
        minimum: 1
        type: integer
      max_users:
        minimum: 1
        type: integer
      name:
        maxLength: 200
        minLength: 2
        type: string
      plan_type:
        $ref: '#/definitions/model.PlanType'
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/organizations:
    get:
      consumes:
      - application/json
      description: Get paginated list of organizations for current user
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number (default 0)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organization.OrganizationListOut'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get organizations
      tags:
      - Organization Management
    post:
      consumes:
      - application/json
      description: Create a new organization
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization data
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/organization.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/organization.OrganizationOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - Organization Management
  /api/v1/organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an organization
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Organization Management
    get:
      consumes:
      - application/json
      description: Get organization details by ID
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organization.OrganizationOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get organization
      tags:
      - Organization Management
    put:
      consumes:
      - application/json
      description: Update organization details
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Organization update data
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/organization.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organization.OrganizationOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - Organization Management
  /api/v1/organizations/{id}/invite:
    post:
      consumes:
      - application/json
      description: Invite a user to join the organization
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Invitation data
        in: body
        name: invite
        required: true
        schema:
          $ref: '#/definitions/organization.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organization.InviteOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Invite member
      tags:
      - Organization Management
  /api/v1/organizations/{id}/invites:
    get:
      consumes:
      - application/json
      description: Get list of pending invitations for the organization
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organization.PendingInvitesOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get pending invites
      tags:
      - Organization Management
  /api/v1/organizations/{id}/members:
    get:
      consumes:
      - application/json
      description: Get list of organization members
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/organization.MembersOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get organization members
      tags:
      - Organization Management
  /api/v1/organizations/{id}/members/{userId}:
    delete:
      consumes:
      - application/json
      description: Remove a member from the organization
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Remove member
      tags:
      - Organization Management
  /api/v1/organizations/{id}/members/{userId}/role:
    put:
      consumes:
      - application/json
      description: Update a member's role in the organization
      parameters:
      - description: Bearer token
        format: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role update data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/organization.UpdateMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/inout.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/inout.BaseResponse'
      security:
      - BearerAuth: []
      summary: Update member role
      tags:
      - Organization Management
...
swagger: "2.0"